requirements.txt
grpcio==1.76.0
grpcio-tools==1.76.0
protobuf==6.33.5
typing_extensions==4.15.0
